Sprint 16. ShareIt. Add Tests
