+++ server/src/main/java/ru/practicum/shareit/constant/Constants.java
@@ -7,4 +7,7 @@ public class Constants {
     public static final String BOOKING_NOT_FOUND_ERR = "Бронирование с id %d не найдено";
     public static final String ITEM_NOT_FOUND_ERR = "Вещь с id %d не найдена";
     public static final String ITEM_REQUEST_NOT_FOUND_ERR = "Запрос с id %d не найден";
+
+    private Constants() {
+    }
 }
